fix: Move profile deployment to a different file
